+++ usr.sbin/trpt/trpt.c
@@ -417,12 +417,12 @@ tcp_trace(short act, short ostate, struct tcpcb *tp, int family __unused,
 	printf("\n");
 	if (sflag) {
 		printf("\trcv_nxt %lx rcv_wnd %lx snd_una %lx snd_nxt %lx snd_max %lx\n",
-		    (u_long)tp->rcv_nxt, tp->rcv_wnd,
+		    (u_long)tp->rcv_nxt, (u_long)tp->rcv_wnd,
 		    (u_long)tp->snd_una, (u_long)tp->snd_nxt,
 		    (u_long)tp->snd_max);
 		printf("\tsnd_wl1 %lx snd_wl2 %lx snd_wnd %lx\n",
 		    (u_long)tp->snd_wl1,
-		    (u_long)tp->snd_wl2, tp->snd_wnd);
+		    (u_long)tp->snd_wl2, (u_long)tp->snd_wnd);
 	}
 	/* print out timers? */
 #if 0
